requirements.txt
python-binance==1.0.35
python-dotenv==1.2.2
typer==0.24.1
